[MIG] partner_multi_relation: Migration to 11.0
* Update README
* Python 2to3
* Move relation menus items under contacts
* Fix date comparison
* Fix cache invalidate on relation write
* Bump module version
* Remove test_relation_type_unlink_dberror

  This test is not required because it tests a standard behavior of SQL.
  If a foreign key column is not nullable, then the foreign row can not be deleted.

* Remove utf8 encoding comment
* Add missing dependency for partner_multi_relation
* Add missing active_test in view action

+++ partner_multi_relation/models/res_partner_relation_type.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 # Copyright 2013-2018 Therp BV <https://therp.nl>.
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl.html).
 """Define the type of relations that can exist between partners."""
@@ -152,7 +151,8 @@ class ResPartnerRelationType(models.Model):
                     # already ended:
                     cutoff_date = fields.Date.today()
                     for relation in invalid_relations:
-                        if relation.date_start >= cutoff_date:
+                        if (relation.date_start and
+                                relation.date_start >= cutoff_date):
                             relation.unlink()
                         elif (not relation.date_end or
                                 relation.date_end > cutoff_date):
